dashboard project loading state
