--- apps/builder/app/dashboard/loading-state/index.tsx
@@ -1,1 +1,0 @@
-export * from "./loading";
