--- apps/builder/app/dashboard/loading-state/svg-loading.tsx
@@ -1,18 +1,0 @@
-import { css } from "@webstudio-is/design-system";
-import svgSpinner from "./svg-spinner";
-
-const svgContainerStyle = css({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-});
-
-export const SvgLoading = ({ size = 64 }: { size?: number }) => {
-  return (
-    <div
-      className={svgContainerStyle()}
-      style={{ height: size, width: size }}
-      dangerouslySetInnerHTML={{ __html: svgSpinner }}
-    />
-  );
-};
