--- apps/builder/app/dashboard/loading-state/svg-spinner.ts
@@ -1,82 +1,0 @@
-const svgSpinner = `
-<svg id="e2CRglijn891" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"
-     viewBox="0 0 128 128" shape-rendering="geometricPrecision" text-rendering="geometricPrecision">
-    <style>
-        <![CDATA[
-          #e2CRglijn892_tr {
-            animation: e2CRglijn892_tr__tr 4000ms linear infinite normal forwards
-          }
-
-          @keyframes e2CRglijn892_tr__tr {
-              0% {
-                  transform: translate(64px, 64px) rotate(90deg);
-                  animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1)
-              }
-              50% {
-                  transform: translate(64px, 64px) rotate(810deg);
-                  animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1)
-              }
-              100% {
-                  transform: translate(64px, 64px) rotate(1530deg)
-              }
-          }
-          
-          #e2CRglijn892 {
-              animation-name: e2CRglijn892_s_p, e2CRglijn892_s_do;
-              animation-duration: 4000ms;
-              animation-fill-mode: forwards;
-              animation-timing-function: linear;
-              animation-direction: normal;
-              animation-iteration-count: infinite;
-          }
-          
-          @keyframes e2CRglijn892_s_p {
-              0% {
-                  stroke: #39fbbb
-              }
-              25% {
-                  stroke: #4a4efa
-              }
-              50% {
-                  stroke: #e63cfe
-              }
-              75% {
-                  stroke: #ffae3c
-              }
-              100% {
-                  stroke: #39fbbb
-              }
-          }
-          
-          @keyframes e2CRglijn892_s_do {
-              0% {
-                  stroke-dashoffset: 251.89
-              }
-              2.5% {
-                  stroke-dashoffset: 263.88;
-                  animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1)
-              }
-              25% {
-                  stroke-dashoffset: 131.945
-              }
-              52.5% {
-                  stroke-dashoffset: 263.88;
-                  animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1)
-              }
-              100% {
-                  stroke-dashoffset: 251.885909
-              }
-              75% {
-                  stroke-dashoffset: 131.945
-              }
-          }
-        ]]>
-    </style>
-    <g id="e2CRglijn892_tr" transform="translate(64,64) rotate(90)">
-        <ellipse id="e2CRglijn892" rx="42" ry="42" transform="scale(-1,1) translate(0,0)" fill="none" stroke="#39fbbb"
-                 stroke-width="16" stroke-linecap="round" stroke-dashoffset="251.89" stroke-dasharray="263.89"/>
-    </g>
-</svg>
-`;
-
-export default svgSpinner;
