+++ apps/builder/app/dashboard/projects/project-card.tsx
@@ -21,9 +21,10 @@ import {
   ShareProjectDialog,
 } from "./project-dialogs";
 import { ThumbnailLink } from "./thumbnail-link";
+import { useNavigation } from "@remix-run/react";
+import { Spinner } from "../spinner";
 
 const containerStyle = css({
-  overflow: "hidden",
   width: theme.spacing[31],
   height: theme.spacing[29],
   borderWidth: 1,
@@ -34,6 +35,11 @@ const containerStyle = css({
   "&:hover, &:focus-within": {
     boxShadow: theme.shadows.brandElevationBig,
   },
+});
+
+const thumbnailStyle = css({
+  position: "relative",
+  overflow: "hidden",
 });
 
 const footerStyle = css({
@@ -174,6 +180,10 @@ export const ProjectCard = ({
   const [isHidden, setIsHidden] = useState(false);
   const { thumbnailRef, handleKeyDown } = useProjectCard();
   const handleDuplicate = useDuplicate(id);
+  const { state, location } = useNavigation();
+  const linkPath = builderPath({ projectId: id });
+  // Transition to the project has started, we may need to show a spinner
+  const isTransitioning = state !== "idle" && linkPath === location.pathname;
 
   return (
     <Box as="article" hidden={isHidden}>
@@ -184,11 +194,11 @@ export const ProjectCard = ({
         tabIndex={0}
         onKeyDown={handleKeyDown}
       >
-        <ThumbnailLink
-          title={title}
-          to={builderPath({ projectId: id })}
-          ref={thumbnailRef}
-        />
+        <Flex className={thumbnailStyle()}>
+          <ThumbnailLink title={title} to={linkPath} ref={thumbnailRef} />
+          {isTransitioning && <Spinner />}
+        </Flex>
+
         <Flex
           justify="between"
           shrink={false}
