+++ apps/builder/app/dashboard/spinner/index.tsx
@@ -1,0 +1,1 @@
+export * from "./spinner";
